Extended StandardCampaign with testable methods.
Made a suggested change to StandardCampaign in order to allow a refund BalanceClaim to be generated by a different user.
This would be in line with how a different user can initiate a payoutToBeneficiary.

Added StandardCampaign tests for payoutToBeneficiary and claimRefundOwedexpiry
while considering various funding goals, expiries, and contribution times.